Move AddToVirtualMachineConfig() to pkg/vf
